% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xval.glm.R
\name{cross.validate}
\alias{cross.validate}
\title{cross validation function (only used internally)}
\usage{
cross.validate(M, folds, n, K, glm.family, data, y, models, loss)
}
\arguments{
\item{M}{NA}

\item{folds}{NA}

\item{n}{NA}

\item{K}{NA}

\item{glm.family}{NA}

\item{data}{NA}

\item{y}{NA}

\item{models}{NA}

\item{loss}{NA}
}
\value{
list with cross-validation output.
}
\description{
cross validation function (only used internally)
}
