% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xval.glm.R
\name{xval.glm}
\alias{xval.glm}
\title{Run X-val on a set of glms}
\usage{
xval.glm(data, models, glm.family = gaussian, folds = 10,
  repeats = 200, loss = NULL, numCore = NULL, plots = T,
  gray = F, seed = NULL)
}
\arguments{
\item{data}{data.frame containing the data}

\item{models}{list of model formulas, each list entry is used as a}

\item{glm.family}{name of the glm family (see \link{glm}() for details), default is gaussian}

\item{folds}{number of folds (default = 10)}

\item{repeats}{number of repeats (default = 200)}

\item{loss}{loss function for the GLM (default = NULL equals RMSE)}

\item{numCore}{number of cores for use with parallel (default = NULL, equals no parallelization)}

\item{plots}{output fancy plots with results}

\item{gray}{output greyscale plots (default = F)}

\item{seed}{seed for the folds (default = NULL, seed will be random)}
}
\value{
object of class \code{xval.glm}
}
\description{
Run X-val on a set of glms
}
