% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xval.glm.R
\name{cross.validate}
\alias{cross.validate}
\title{cross validation function (only used internally)}
\usage{
cross.validate(M, folds, n, K, glm.family, data, y, models, loss)
}
\description{
cross validation function (only used internally)
}
